+++ exp/select_clauses_test.go
@@ -393,6 +393,30 @@ func (scs *selectClausesSuite) TestGroupBy() {
 	scs.Equal(g, c2.GroupBy())
 }
 
+func (scs *selectClausesSuite) TestGroupByAppend() {
+	g := exp.NewColumnListExpression(exp.NewIdentifierExpression("", "", "a"))
+	g2 := exp.NewColumnListExpression(exp.NewIdentifierExpression("", "", "b"))
+
+	c := exp.NewSelectClauses().SetGroupBy(g)
+	c2 := c.GroupByAppend(g2)
+
+	scs.Equal(g, c.GroupBy())
+
+	scs.Equal(exp.NewColumnListExpression(g, g2), c2.GroupBy())
+}
+
+func (scs *selectClausesSuite) TestGroupByAppend_NoPreviousGroupBy() {
+	g := exp.NewColumnListExpression(exp.NewIdentifierExpression("", "", "a"))
+	g2 := exp.NewColumnListExpression(exp.NewIdentifierExpression("", "", "b"))
+
+	c := exp.NewSelectClauses().GroupByAppend(g)
+	c2 := c.GroupByAppend(g2)
+
+	scs.Equal(g, c.GroupBy())
+
+	scs.Equal(exp.NewColumnListExpression(g, g2), c2.GroupBy())
+}
+
 func (scs *selectClausesSuite) TestSetGroupBy() {
 	g := exp.NewColumnListExpression(exp.NewIdentifierExpression("", "", "a"))
 	g2 := exp.NewColumnListExpression(exp.NewIdentifierExpression("", "", "b"))