+++ sqlgen/select_sql_generator.go
@@ -15,7 +15,7 @@ type (
 	}
 	// The default adapter. This class should be used when building a new adapter. When creating a new adapter you can
 	// either override methods, or more typically update default values.
-	// See (github.com/doug-martin/goqu/adapters/postgres)
+	// See (github.com/doug-martin/goqu/dialect/postgres)
 	selectSQLGenerator struct {
 		CommonSQLGenerator
 	}