+++ update_dataset.go
@@ -10,7 +10,7 @@ import (
 type UpdateDataset struct {
 	dialect      SQLDialect
 	clauses      exp.UpdateClauses
-	isPrepared   bool
+	isPrepared   prepared
 	queryFactory exec.QueryFactory
 	err          error
 }
@@ -35,12 +35,12 @@ func Update(table interface{}) *UpdateDataset {
 // prepared: If true the dataset WILL NOT interpolate the parameters.
 func (ud *UpdateDataset) Prepared(prepared bool) *UpdateDataset {
 	ret := ud.copy(ud.clauses)
-	ret.isPrepared = prepared
+	ret.isPrepared = preparedFromBool(prepared)
 	return ret
 }
 
 func (ud *UpdateDataset) IsPrepared() bool {
-	return ud.isPrepared
+	return ud.isPrepared.Bool()
 }
 
 // Sets the adapter used to serialize values and create the SQL statement
@@ -236,7 +236,7 @@ func (ud *UpdateDataset) Executor() exec.QueryExecutor {
 }
 
 func (ud *UpdateDataset) updateSQLBuilder() sb.SQLBuilder {
-	buf := sb.NewSQLBuilder(ud.isPrepared)
+	buf := sb.NewSQLBuilder(ud.isPrepared.Bool())
 	if ud.err != nil {
 		return buf.SetError(ud.err)
 	}